tools/augnet/requirements-diff.txt
# Phase 3 differential validation environment (Python 3.10 recommended)
# Pinned for deterministic music21 baseline export and parity diff checks.
music21==9.1.0
onnxruntime==1.17.1
numpy==1.24.3
pytest==8.2.2
